+++ packages/plugin-hardhat/src/defender/deploy.ts
@@ -19,8 +19,8 @@ import UpgradeableBeacon from '@openzeppelin/upgrades-core/artifacts/@openzeppel
 import TransparentUpgradeableProxy from '@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/transparent/TransparentUpgradeableProxy.sol/TransparentUpgradeableProxy.json';
 import ProxyAdmin from '@openzeppelin/upgrades-core/artifacts/@openzeppelin/contracts/proxy/transparent/ProxyAdmin.sol/ProxyAdmin.json';
 
-import { getNetwork, getDeployClient } from './utils';
-import { DeployTransaction, DefenderDeployOptions, UpgradeOptions } from '../utils';
+import { getNetwork, getDeployClient, parseTxOverrides } from './utils';
+import { DeployTransaction, DefenderDeployOptions, UpgradeOptions, EthersDeployOptions } from '../utils';
 import debug from '../utils/debug';
 import { getDeployData } from '../utils/deploy-impl';
 import { ContractSourceNotFoundError } from '@openzeppelin/upgrades-core';
@@ -57,7 +57,7 @@ type CompilerOutputWithMetadata = CompilerOutputContract & {
 export async function defenderDeploy(
   hre: HardhatRuntimeEnvironment,
   factory: ContractFactory,
-  opts: UpgradeOptions & DefenderDeployOptions,
+  opts: UpgradeOptions & EthersDeployOptions & DefenderDeployOptions,
   ...args: unknown[]
 ): Promise<Required<Deployment & RemoteDeploymentId> & DeployTransaction> {
   const client = getDeployClient(hre);
@@ -91,6 +91,7 @@ export async function defenderDeploy(
     relayerId: opts.relayerId,
     salt: opts.salt,
     createFactoryAddress: opts.createFactoryAddress,
+    txOverrides: parseTxOverrides(opts.txOverrides),
   };
 
   let deploymentResponse: DeploymentResponse;